Move validate_field and error methods into resource init
